refactor: remove dead code and contextual panel (~5000 lines)
Delete 13 unused files totaling ~188KB:
- Contextual panel: app/contextual/page.tsx, contextual-operations-panel.tsx,
  contextual-context-canvas.tsx, contextual-diff-explorer.tsx, lib/contextual-chat.ts
- Dead components: chat-input.tsx (73KB, replaced by chat-input-simple.tsx),
  assistant-workflow-board.tsx, create-sweep-dialog.tsx, left-panel.tsx,
  sweep-form-popover.tsx, history-panel.tsx, chat-view.tsx, artifacts-panel.tsx

Fix imports:
- Update ChatMode type imports to use chat-input-simple (canonical source)
- Remove contextual route from handleTabChange in page.tsx
